# perf(core-database-postgres): add index on transactions (timestamp)