--- a/packages/core-database-postgres/lib/migrations/20181204200000-transactions-timestamp-index.sql
+++ b/packages/core-database-postgres/lib/migrations/20181204200000-transactions-timestamp-index.sql
@@ -1,0 +1,1 @@
+CREATE INDEX IF NOT EXISTS "transactions_timestamp" ON ${schema~}.transactions ("timestamp");