--- a/packages/core-database-postgres/lib/migrations/index.js
+++ b/packages/core-database-postgres/lib/migrations/index.js
@@ -7,4 +7,5 @@ module.exports = [
   loadQueryFile(__dirname, './20180305400000-create-transactions-table.sql'),
   loadQueryFile(__dirname, './20181129400000-add-block_id-index-to-transactions-table.sql'),
   loadQueryFile(__dirname, './20181204100000-blocks-generator_public_key-index.sql'),
+  loadQueryFile(__dirname, './20181204200000-transactions-timestamp-index.sql'),
 ]